feat: add ability to generate `patch` files based on FixIt Hints
This change introduces `createPatch()`, which allows you to create
a unix `patch` file-string based on the Fix-It Hints associated with the
`DiagnosticsMessage`.

This can be useful in case you want to generate your own `patch`-files
and use the `patch`-cli tool to apply your patches.

If you want to directly apply the `patch` to the associated file,
then simply use `DiagnosticsMessage.applyFixitHints()` (as was
already exposed in the previous version).